requirements.txt
alembic==1.4.2
amqp==2.6.1
billiard==3.6.3.0
blinker==1.4
cachelib==0.1.1
celery==4.4.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.0
Flask==1.1.2
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Session==0.3.2
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
Flask-WTF==0.14.3
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
Mako==1.1.2
MarkupSafe==1.1.1
mypy==0.782
mypy-extensions==0.4.3
psycopg2==2.8.5
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
rope==0.17.0
six==1.14.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
uWSGI==2.0.18
vine==1.3.0
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
zope.event==4.4
zope.interface==5.1.0
